requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
google-generativeai==0.6.0
requests==2.32.3
Flask-Cors==4.0.1
Flask-Cors
numpy
Flask-APScheduler
